Sketching out a Samza integration
New plan: Samza gets deployed as a) a tar with a bunch of jars /
scripts, and b) a config file. The user's compiled code should be able
to take the graph and generate a bunch of configs, which include
serialized functions to run -- individual jobs can then be launched with
both the tar and the config files.