Exactly once merges
This adds a coast-specific Kafka-backed system as well, which is
responsible for ensuring that messages are sent in the right order.

The merge implementation is rather heavyweight, since it involves
writing all the data twice. (It turns out to be hard to do the
lightweight version with the existing interfaces.) This should be enough
for a proof of concept, but it will need to be revisited later.