Multiple samza job compilers
Living the dream here... `coast` now has a safe and a dangerous mode.

Not too pleased about some of the hacks going on, but I'm also not
interested in burning a bunch of time on it when I already plan to merge
things again later. Thankfully, I don't think it makes things that much
more confusing for the 'safe' case.