Change partitioning strategy
There are two use-cases for the partitioner I'm particularly interested
in:

- Locality-sensitive hashing: this puts 'similar' data in the same
  partition, which sometimes helps you avoid extra network traffic.
- Kafka's default partitioning: `coast` tries to play nicely with
  existing tools, so it's important that it can use the same
  partitioning strategies as Kafka out-of-the-box.

In theory, the last implementation could support either, depending on how
you interpreted the HashCode. The issue was that they didn't mix well --
if you had some old-school topics and some locality-sensitive, you'd
need to configure different 'interpretations' for different topics,
which is pretty error-prone. Instead of that mess, the partitioner now
makes all the decisions -- though that requires passing a little extra
data around.