Update the Twitter Reach example
More info!

Also restricting the type of arguments to `flatten`, and adding a new
`flattenOption` method. The old `flatten` method would flatten streams
of anything, including `Set`s or other unordered containers, which is
not good -- this is a bit gross, but it's at least correct!

I messed around a bit with defining a Sequential typeclass; I'll
probably come back to this later, once the API's solidified a bit.