Allow cycles in the graph
This appears to be well-defined, so I might as well allow it. (The
syntax discourages doing it too often, which I think is the right call.)

To get nontrivial cycles (with one or more intermediate streams) you
need to be able to add name bindings *within* the definition of a cycle.
The `Graphable` trait makes this possible without introducing extra
syntax in the common case.